requirements.txt
Django==5.0.10
djangorestframework==3.15.2
django-htmx==1.19.0
django-environ==0.13.0
celery==5.3.6
redis==5.0.3
Pillow==10.3.0
reportlab==4.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.1.0
django-cors-headers==4.4.0
